abstracted side control buttons
